+++ ChangeLog.md
@@ -12,6 +12,7 @@ Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
 ### Added
  * Declare that the `renderUIFragment()` function accepts undefined input, too
  * Added `NavRadioButtonBarAnt` component, which is just like the `NavMenuBarAnt`, but with radio buttons 
+ * Adding `PollingUpdaterAnt` component, for displaying some content that is periodically updated
 
 ### Fixed
  * Fixed a TS warning around btoa-lite